fix: add  list layer to page
