+++ src/pages/Departments/index.ts
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export const DepartmentsPage = lazy(() => import("./Departments.page"));
