+++ src/pages/skills/index.ts
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export const SkillsPage = lazy(() => import("./Skills.page"));
